% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr10}
\alias{chr10}
\title{Chromosome 10 of T47D}
\format{
The third column refers to rounded log count
}
\usage{
data(chr10)
}
\description{
Intra region of chromosome 10 in cell line T47D (500kb resolution)
}
\keyword{datasets}
