% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr10_20}
\alias{chr10_20}
\title{Chromosome 10_20 of T47D}
\format{
The third column refers to rounded log count
}
\usage{
data(chr10)
}
\description{
Inter region between chromosome 10 and 20 in cell line T47D (500kb resolution)
}
\keyword{datasets}
