% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr20}
\alias{chr20}
\title{Chromosome 20 of T47D}
\format{
The third column refers to rounded log count
}
\usage{
data(chr20)
}
\description{
Intra region of chromosome 20 in cell line T47D (500kb resolution)
}
\keyword{datasets}
