% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranScan.R
\name{TranScan}
\alias{TranScan}
\title{Hi-C inter-chromosomal translocation detection via Scan clustering}
\usage{
TranScan(
  chr1,
  chr2,
  chr1_2,
  outputpath,
  sizes,
  resolution = 5e+05,
  quantitest = 256,
  alpha = 0.05,
  gammath = 0.1,
  number_of_bandwidth = 20,
  diag_remove = 2e+06,
  add_gap = 2500000
)
}
\arguments{
\item{chr1}{First chromosome}

\item{chr2}{Second chromosome}

\item{chr1_2}{Inter-chromosomal region of chr1 and chr2}

\item{outputpath}{The path to store all results}

\item{sizes}{Chrom sizes of chr1_2; vector of length 2}

\item{resolution}{Resolution of Hi-C data}

\item{quantitest}{Grid of testing, must be 2^k, default 256, meaning 256 * 256}

\item{alpha}{FDX, default 0.05}

\item{gammath}{Gamma threshold, default 0.1}

\item{number_of_bandwidth}{Number of bandwidths used, default 20}

\item{diag_remove}{Removed short-range interactions, must be k * resolution, default 2e+06}

\item{add_gap}{Added gap between two chromosomes, must be k * resolution, defult 2500000}
}
\value{
A heatmap and a rejected region plot will be generated under the output folder. The matrix form of rejected region is saved as RR.txt. All intermediate objects will also be saved.
}
\description{
This function allows you to detect the inter-chromosomal translocation
from Hi-C data. It requires two intra-regions and the corresponding
inter-region as inputs. We suggest rounded log counts as inputs.
}
\examples{
data(chr10)
data(chr20)
data(chr10_20)
p_sizes = "../hg38.sizes"
p_outdir = "output"
sizes = read.table(p_sizes, col.name=c("chrom", "size"))
chr10_l = subset(sizes, chrom=="chr10")$size
chr20_l = subset(sizes, chrom=="chr20")$size
size_dim = c(chr10_l, chr20_l)
TranScan(chr10, chr20, chr10_20, p_outdir, size_dim)
}
